AI NEW MODEL ROTATION YO

+++ main/java/net/minecraft/client/renderer/entity/RenderLightningBolt.java
@@ -33,8 +33,8 @@ public class RenderLightningBolt extends Render<EntityLightningBolt>
         {
             adouble[i] = d0;
             adouble1[i] = d1;
-            d0 += (double)(random.nextInt(11) - 5);
-            d1 += (double)(random.nextInt(11) - 5);
+            d0 += (random.nextInt(11) - 5);
+            d1 += (random.nextInt(11) - 5);
         }
 
         for (int k1 = 0; k1 < 4; ++k1)
